Add media query 700px - 990px
In this size of viewport, the video section had a narrow width, the font
size of the titles of prices were large, and the the prices were thin.
Therefore, the size of the video section increased, the font size of the
prices decreased, and the made the .price divs thicker.